Moved based on namespace change
