+++ tests/EyeChart/Tests/VO/Authentication/AuthenticationVOTest.php
@@ -7,10 +7,12 @@ declare(strict_types=1);
  * (c) 2017
  */
 
-namespace EyeChart\Tests\VO;
+namespace EyeChart\Tests\VO\Authentication;
 
 use EyeChart\Mappers\AuthenticateMapper;
-use EyeChart\VO\AuthenticationVO;
+use EyeChart\Tests\VO\VOTest;
+use EyeChart\VO\Authentication\AuthenticationVO;
+use EyeChart\VO\Authentication\CredentialsVO;
 
 /**
  * Class AuthenticationVOTest
@@ -21,13 +23,13 @@ final class AuthenticationVOTest extends VOTest
     /** @var string */
     protected static $subjectVONameSpace = AuthenticationVO::class;
 
-    /** @var array[]  */
+    /** @var array[] */
     protected static $validAssertionValues = [
         'username' => ['kirk'],
         'password' => ['000 Destruct 0'],
     ];
 
-    /** @var array[]  */
+    /** @var array[] */
     protected static $invalidAssertionValues = [
         'username' => [''],
         'password' => ['']
@@ -39,6 +41,9 @@ final class AuthenticationVOTest extends VOTest
 
         self::$invalidAssertionValues['token'][] = str_repeat('a', AuthenticateMapper::TOKEN_LENGTH + 1);
         self::$invalidAssertionValues['token'][] = str_repeat('a', AuthenticateMapper::TOKEN_LENGTH - 1);
+
+        self::$validAssertionValues['derivedCredentials'][] = new CredentialsVO();
+        self::$validAssertionValues['storedCredentials'][]  = new CredentialsVO();
 
         parent::$subjectVONameSpace     = self::$subjectVONameSpace;
         parent::$validAssertionValues   = self::$validAssertionValues;
